Rename `types.py` to `aliases.py`
