+++ tests/row_generation/generator_test_helpers.py
@@ -1,7 +1,7 @@
 from typing import List, Iterator
 
 from wheatley.bell import Bell
-from wheatley.types import Stroke, HANDSTROKE
+from wheatley.aliases import Stroke, HANDSTROKE
 
 
 def rounds(stage: int) -> List[int]:
